reinterpret_cast

+++ cpp/open3d/t/io/NumpyIO.cpp
@@ -521,8 +521,8 @@ std::tuple<uint16_t, size_t, size_t> ParseZipFooter(FILE* fp) {
     uint16_t disk_start           = *reinterpret_cast<uint16_t*>(&footer[6]);
     uint16_t nrecs_on_disk        = *reinterpret_cast<uint16_t*>(&footer[8]);
     uint16_t nrecs                = *reinterpret_cast<uint16_t*>(&footer[10]);
-    size_t   global_header_size   = *reinterpret_cast<uint32_t*>(&footer[12]);
-    size_t   global_header_offset = *reinterpret_cast<uint32_t*>(&footer[16]);
+    uint32_t global_header_size   = *reinterpret_cast<uint32_t*>(&footer[12]);
+    uint32_t global_header_offset = *reinterpret_cast<uint32_t*>(&footer[16]);
     uint16_t comment_len          = *reinterpret_cast<uint16_t*>(&footer[20]);
     // clang-format on
